kafka_script/requirements.txt
faker
python-dotenv
kafka-python 